+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Pixel events of `addToCart` and `removeFromCart`.
+
 ## [2.23.0] - 2019-07-18
 
 ### Changed